Module and other fixes
Replaced subprocesses in main.py with calls to [MODULE].main().

Added raw sequence attribute to Feature class.

Fixed a bug in which a BLAST query with the same display name as a BLAST hit would double-count the number(s) on the tspan ID for that line/residue.
Now uses full protein names (those in the original .clustal alignment) in source XML IDs in order to avoid this.